+++ src/easing/index.js
@@ -4,6 +4,8 @@ import { cubic as cubic_ } from './cubic'
 import { sine as sine_ } from './sine'
 import { exponential as exponential_ } from './exponential'
 import { elastic as elastic_ } from './elastic'
+import { circle as circle_ } from './circle'
+import { back as back_ } from './back'
 import { withDomainChecker } from './easingUtils'
 
 const linear = withDomainChecker(linear_)
@@ -12,6 +14,8 @@ const cubic = withDomainChecker(cubic_)
 const sine = withDomainChecker(sine_)
 const exponential = withDomainChecker(exponential_)
 const elastic = withDomainChecker(elastic_)
+const circle = withDomainChecker(circle_)
+const back = withDomainChecker(back_)
 
 export {
   linear,
@@ -20,4 +24,6 @@ export {
   sine,
   exponential,
   elastic,
+  circle,
+  back,
 }